fix: hong kong is not in china
a previous commit set this conditional to return as `true` when
the detected windows locale `zh-hk` is detected, and then applies
`GOPROXY` setting based on such detection.

the flag is supposed to _Set GOPROXY for people in China_. if this
flag is truthy, the proxy `goproxy.cn` is used, which would, in
effect, redirect HK users traffic.